pyproject.toml
[project]
name = "unitycatalog-pydantic"
version = "0.1.0"
description = "Support for Unity Catalog functions in Pydantic"
authors = [{name="Daniël Tom", email="detom89@gmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2",
    "unitycatalog-ai>=0.1",
    "unitycatalog-client>=0.2",
]

[project.scripts]
unitycatalog-pydantic = "unitycatalog_pydantic:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "delta-spark>=3.2.1",
    "polyfactory>=2.18.1",
    "pyspark>=3.5.4",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.21.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.4",
    "setuptools>=75.6.0",
    "testcontainers>=4.9.0",
]
